doc: cmake: Add makeinfo wrapper
Instead of different flavors of `texi2???` one might as well use
`makeinfo`. This also removes the dependency on `texinfo.tex` being
present in the build directory.

This was appearently the __wrong__ one from `automake` and not the one
provided by `texlive`.

`*.texi` files are just copied for the time being.

Signed-off-by: brutzzl3r <s3b.gr0ss@gmail.com>