doc: cmake: Generate texi files
Move that obfuscated `awk` | `sed` magic into the script and try to
bring some light on it.
At least put it somewhere where no one needs to see it.
However it works - Compared the output to the files generated by `make`
and beside of some diffs due to the wrong `avrdude` was used it looks
fine.

Signed-off-by: brutzzl3r <s3b.gr0ss@gmail.com>